+++ src/main/java/com/example/demo/repository/DrinkRecordDao.java
@@ -1,5 +1,6 @@
 package com.example.demo.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.example.demo.entity.DrinkRecord;
@@ -8,6 +9,6 @@ public interface DrinkRecordDao {
 	
 	List<DrinkRecord> list();
 	
-	long drinkSum();
+	BigDecimal drinkSum();
 
 }
