+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE alcohol_type (
   type_id int(2) NOT NULL,
   name varchar(30) NOT NULL,
-  volume int(2) NOT NULL,
+  volume decimal(2) NOT NULL,
   PRIMARY KEY (type_id)
 );
 
@@ -9,7 +9,7 @@ CREATE TABLE drink_record (
   id int(5) NOT NULL AUTO_INCREMENT,
   user_id int(5) NOT NULL,
   type_id int(2) NOT NULL,
-  quantity int(3) NOT NULL,
+  quantity decimal(3) NOT NULL,
   date datetime NOT NULL,
   PRIMARY KEY (id)
 ) ;
@@ -19,6 +19,6 @@ CREATE TABLE user (
   username varchar(50) NOT NULL,
   email varchar(70) NOT NULL,
   password varchar(60) NOT NULL,
-  weight int(3) NOT NULL,
+  weight decimal(3) NOT NULL,
   PRIMARY KEY (user_id)
 );
